Trying to get rid of some of the weird dependencies from the last UI fix commit. For example, WorldPinOnMapController depended on an platform interiors type when the only information it need to know was whether or not we're in interiors mode.
Added IAppModeModel to try and remedy this. At the moment, it's just observing interiors and setting whether or not we're in interior mode (and is only used by the above controller). Ideally, this might observe a few systems and toggle modes accordingly (for example, observes modality model to tell whether we're in a menu or not). Even more ideally, it might be some controller type, whereby we can also set what mode we want to be in and it handles all of the internal setting. Currently, the interiors makes this a little difficult, as we enter using a TryEnterInterior API point that returns a callback if successful.

In terms of usage, I'm only relying on this for interiors at the moment. So, for example, the controller that uses it still checks the modality model separately.

Buddy: Mark