Add more error-checking to display binding.
This is part of MPLY-6280. Currently, we've got a few crashes in Recce
where a builder on a thread cannot set the context. I believe the
context is always invalid in this case, just we weren't checking the
error codes during creation. Won't fix the bug, but it will blow up at a
well-known location rather than at some point down the line.

Buddy: Vimmy & Scott M