Android fix for MPLY-7122.
Added a new menu_list_subitem_with_details view. It’s an amalgamation
of the subitem and search item views (has the sizes of subitem but some
of the properties, such as two text views, of the search item).

Might seem a little DRY fail, but prefer this approach to programmatic
switching in a single view (see the previous iOS commit).

MenuItemData now stores a details string, which it parses from the
serialised JSON it receives.

The adapter checks to see if we have details on getChildView and
inflates whichever view it needs accordingly, assigning text to the
details field if necessary.

Buddy: Judith