MPLY-8448: Limit Senion broadcast, buddy: Jonty D
When bluetooth was disabled, `didLocationError` was only checking if
there were errors before broadcasting. Now it checks to see if bluetooth
is active AND free of errors before broadcasting.

`bluetoothAdapter != null` and `bluetoothAdapter.isEnabled()` may be
expensive operations, and might need to rely on updating status on
change instead of being checked every time.