Update to iOS 9
Changes to deprecated methods from iOS 8 to 9.  The removal of UIAlertView is causing a lot of work here as the replacement UIAlertAction should only be called from a UIVIewController and many of the classes using UIAlertView are plain UIViews.
The string escape for % encoding has changed.
Removal of UIPopoverController.  This is replaced with UIModalPresentationPopover to make it consistent between iPhone and iPad, so references to it can be removed as the iPhone version will now cover both.