MPLY-7158: Android pin description, buddy: Michael C
Android `MyPinDetailsView` class has a `display` function with logic for
hiding empty description fields when displaying pins. This logic was
never hiding the description fields because when the user's description
was being read it would be given a default value of `"No description"`.

Now `descriptionText` is not given the value of `"No description"` if
empty, and so the logic in `MyPinDetailsView` will hide empty
descriptions.