MPLY-8728: "pin" replaces "misc", buddy: Mark S
The `misc` tag is to be removed, as it is a synonym to the `pin` tag,
and so all direct references within the app code to `misc` have been
replaced with references to the `pin` tag.

These changes affect very little, but prevents the removal of the `misc`
tag from breaking functionality.