Fix for scroll button sometimes not disappearing on iOS. Buddy: Paul
For some reason the self.pSearchResultsTableView.contentOffset.y value
wasn’t always the same when you ran the app. Couldn’t see an obvious
reason why this would be, so removed == test for whether to show button.
Also flipped the logic of shouldShowScrollButton function so it returns
true when it should show and false when it shouldn’t.