MPLY-9103: Check for empty yelp map, buddy: Sam A
If no custom yelp mapping is defined in the standard_config.json then
the yelp mapping is set to "unused_string".

When checking if the default yelp mapping should be overwritten by any
custom mapping, the check was that the string was longer than 0
characters. The string would have been set to "unused_string" which is
longer than 0 characters, and would overwrite the default yelp mapping
with "unused_string".

Now the check is that the string is longer than 0 characters AND is not
"unused_string".