MPLY-9129 fix, Currently selected pin indicator has disappeared from Swallow. Buddy: Andrew
Ignore occlusion had been set to false for the selected label styles on
iOS which meant the highlight was being hidden as it appears behind the
selected pin. The priority for the selected label styles on Android and
Windows has been changed to match iOS to fix the selected label
overlapping other labels with priority 0.