Adding timer for Senion disconnect. Buddy: John
When walking between north and south buildings in 3QVS, Senion
disconnects as the north and south buildings have different maps in
Senion. Previously we immediately switched back to gps which kicked you
out the building, now we use a timer so we don’t switch to gps till
after an amount of time meaning if you disconnect and then reconnect
quickly you won’t be kicked out the building.