MPLY-9120 fix on Android, Running multiple searches without closing the LHS results... Buddy: Judith
There was a delayed call to close the search results section when no
results were returned which was added to allow any visible search items
to animate out before the search menu results section closed. This
caused MPLY-9120 because a search cancels any current search which
updates the search menu with 0 results and because the call to clear
the search results section was delayed, the search results would come
in and then the search results section would be closed.