Add network SSID timeout, buddy: David C
Previously users were kicked out of a building that required them to be
on a trusted network when they changed from one trusted network to
another trusted network.

Now the app has a sixty second grace period that starts when one network
disconnects, and ends by kicking the user out of the building unless
they have connected to a trusted network within that time period.