Tentative fix for MPLY-9055
The wiring of this stuff is hard to understand and could do with a
clean-up / refactor.

WorldPinsService is tracking the selected pin by using at least 3 pieces
of state
 - m_selectedSearchResultId (the unique identifer that is stable across
 searches, e.g. "395595"
 - m_pSelectedPinHighlight (ptr to a highlight pin)
 - m_selectedPinId (transient pin id that is not stable across searches)

When you re-search (even using the same term), it will call RemovePin
for each visible pin, then re-add them using AddPin. This changes the
transient pin ids, and causes m_pSelectedPinHighlight to no longer match
(as the highlight pin is destroyed when RemovePin() is called).
Similarly, m_selectedPinId cannot refer to the old one. The 3 bits
of state go out of sync.

Bodged in a fix by ensuring m_selectedSearchResultId is not cleared when
the re-search occurs, and we can then find & re-add the highlight using
the stable id.

Not a great fix, but already spent too long on this.

Buddy: Malc B