Search widget integration: remove JSON, return an array of structures
Search results are no longer serialized/deserialized.
Results come in as an array of structures from the native thread,
and these are now passed directly to the search provider.
In the Android implementation, the results need to be translated
to an array of Java structures, via JNI calls.

Buddy: David Crooks