Search widget integration: search result selection
No longer using the DefaultSearchResult, but a custom one
(SearchWidgetResult) including an index into the array
of results. The SearchServicesResult returned by the
SearchService is including the index into the original
array of results (as coming from the SDK thread) and
a copy of the SDK result structure.

Buggy version! - the SDK thread refreshes the search results
(when entering an interior, or when scrolling the map)
without the UI widget taking any notice. Fixing this requires
changes in the SDK thread (in a future commit).

Buddy: Andrew Gurney