Bidirectional bus: debugging Eegeo_TTY (between #ifdef/#endif)
Eegeo_TTY on Publish, to debug the message flow.

Disabled (no #define), so no effect to anyone.
If enabled for debugging: Eegeo_TTY does not appear to be thread-safe,
so you're on your own. ALWAYS disable for production (never push it
enabled to a repository).

Buddies: Jean Grizet, Mark Simpson