Search widget integration: replace query refreshing by a message to the UI
Refreshing query results should originate from the new search widget;
before, the Native thread was re-performing the search on its own
(upon entering an interior, or when scrolling the map),
leading to the widget list of results getting out of sync.

In this commit, the Native thread does not re-query results any more;
instead, it published a message to the UI thread requesting a refresh
of the query. The UI-thread side (listening to this message, and
actually re-doing the query) will be added in a future commit.

Buddy: Jordan Brown