Search widget integration: sensible parameters to the search refresh message
Upon a search results refresh, a message is now sent to the
UI thread, so that the refresh comes from the search widget down.
The parameters of this message were placeholders, and now are
a more sensible query object, location and radius.

The message is still to be listened to and interpreted in the
UI thread (to be done in a future commit).

Buddy: David Crooks