Search widget integration: refresh search results
The view controller passes the messages to the view and this to the
widget, which does a new search (with context) via its provider.

The message going back into the SDK thread does not yet has all
context parameters required (the existing message did not have
location and radius). A new message will be added in a future commit.

Buddy: Jean Grizet