Search widget integration: (tag search) menu item text passed to the controller
When a (tagged search) menu item is pressed, the text of this item
is passed from the view to the controller. The controller uses a
std::map to transform this text into the appropriate search tag.
(The actual search is not implemented yet - will come in the next
commit.)

Buddy: David Crooks