Search widget integration: blank display text in widget on tag searches
A new entry point in the search widget was provided (by Michael
Odonnell), where a display text can be shown (previously the
query string and the display text were the same thing).
This new method is used to do tag searches while keeping
the edit text in the widget blank (as opposed to showing
the tag, which was ugly).

Buddy: David Crooks