Search widget integration: separate class for tags on the UI thread
This is part of the (quick&dirty) modification to include radius
override on tag searches. This commit just refactors all the tag
collection into a separate class, and deals with the messages that load
the collection. The actual radius overriding will be done in a next
commit.

Buddy: Tom Paynter