Search widget integration: hide Discover menu on interiors, part 1
Not really hiding the Discover menu yet in Android, because of recent
modifications (by someone else) in the way menu entries are
loaded/refreshed. A next commit (by someone else) will fix this.

There is a quick&dirty change on the iOS side, just to prevent it from
crashing. Incidentally, iOS does hide the Discover menu on interiors.
This code will be cleaned up in a next commit (by me).

Buddy: David Crooks