Fix to MPLY-9465: decouple interior elevator slider from floor index
Only a maximum number of floors is shown in the slider; if the building
has more floors than shown, the slider tags change when you move the slider
against the top or bottom (as it does on Android, iOS or the web).

The controller implements now IUpdateableViewController and receives
frame update events (and passes them to the WPF view). This was needed
in order to change floor labels even when the slider is pressed against
one end; previously, the view received only slider change events, and
there was no way of scrolling floors tags when the slider is against
one end (and hence not changing).

Buddy: Andrew Gurney