Additional fix to MPLY-9465 (indoor elevator): properly settle on a floor
At the ends of the slider, an "update frame" event needs to be used,
because the slider doesn't move when pressed against one of the ends
(no chance for drag move events to happen if the thumb doesn't move).
In this commit, the "update frame" event was changed in order to
have an effect only when a drag is in process; this way the
"drag complete" event will settle on a floor.

Buddy: Jean Grizet