Ignore Senion bluetooth disconnect, buddy: David C
Prior to this commit, the app would start the blue sphere timeout as
soon as Senion lost connection to bluetooth (notified by
SSIStepInsideSdkErrorTypeBleNotEnabled being passed to an error handling
function), but Senion would still have a location as it uses motion
sensors in conjunction with bluetooth.

Now we wait until Senion reports that it no longer has a location
available before initiating the timeout.

Other than removing lines 256/257 from SenionLabLocationManager.mm, the
other changes are to clarify that we are checking if Senion has a
location available, rather than if Senion is connected to bluetooth.