Widget integration: tag searches and search refreshes now working
Query message received in the native thread and passed to the
SearchQueryPerformer. Also, the SearchRefreshService does not
initiate a search directly, but sends a message to the UI thread
for the widget to initiate the search refresh.

Tags/refreshes no longer have a "radius override", as they did in Swallow.

Buddy: David Crooks