DEL-42: removing SQLite database, step 1: removing SwallowPoiDbModule
The module makes use of an SQLite database, and Tim tells me
it can be removed. Occurrences of the actual SQLite database
(the variable, abd the method used to open or create the database)
were also removed from MobileExampleApp.

At this point the app does not compile, but I want to do this
big change in steps (on a Git branch).

Buddy: John Bell