Fix for the compass button going off screen and never coming back
A previous fix for MPLY-9710 was reverted (meaning that the app
will crash now on startup queries). The fix was incorrect and
caused the compass to go away forever.

MPLY-9710 is ultimately caused by the event handler at
SearchWidgetController::OnViewClose being reentered
(the initial conditions on a startup query are such
that this handler triggers a second Close event).
I am working on a fix, but wish to clear the most urgent
compass issue first.