Fix to MPLY-9710: crash on startup queries
Performing a query upon start (or on a deeplink when the app
was not running) set the app into a 'wrong' state: a query was
active but the search result view was deemed open. In this 'wrong'
initial state, the reactor logic would reenter the event handler
SearchWidgetController::OnViewClose (the handler would trigger
a second close event), which failed an assertion.

Fixed by closing the menu (which closes the results view)
on startup queries or deeplinks.

Buddy: David Crooks